# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/review": {
      "post": {
        "operationId": "OrdersController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/approve": {
      "post": {
        "operationId": "OrdersController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/reject": {
      "post": {
        "operationId": "OrdersController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "session": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Session"
                        }
                      ],
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/recover": {
      "post": {
        "operationId": "AuthController_recover",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      },
      "patch": {
        "operationId": "AuthController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Apollo Rigor API",
    "description": "Vendas e locações avulsas. Valores em centavos; datas YYYY-MM-DD. Checkout exclusivamente simulado. Auth intermediado pelo NestJS/Supabase.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SALE",
              "RENTAL"
            ]
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "variantId",
          "type"
        ]
      },
      "RejectOrderDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 1000,
            "pattern": "\\S"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "pattern": "\\S"
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RecoverDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "password"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "pattern": "\\S"
          },
          "phone": {
            "type": "string",
            "minLength": 8,
            "maxLength": 25
          },
          "document": {
            "type": "string",
            "pattern": "^(\\d{11}|\\d{14})$"
          }
        }
      },
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "authUserId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "document": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "CLIENT",
              "ADMIN"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Variant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "size": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "collection": {
            "type": "string"
          },
          "fabric": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "rentalPriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "salePriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            }
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "transactionId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "simulated": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "paidAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SALE",
              "RENTAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "CONFIRMED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "priceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pickedUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "damageNotes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "payment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Payment"
              }
            ],
            "nullable": true
          },
          "variant": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Variant"
              }
            ],
            "properties": {
              "product": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        }
      },
      "OrderHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "orderId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "protocol": {
            "type": "string"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "variantId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SALE",
              "RENTAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "UNDER_REVIEW",
              "APPROVED",
              "REJECTED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-10-15"
          },
          "quotedPriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "notes": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "customerDocument": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "history": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderHistory"
            }
          },
          "transaction": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Transaction"
              }
            ],
            "nullable": true
          },
          "variant": {
            "$ref": "#/components/schemas/Variant"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "integer"
          },
          "expiresIn": {
            "type": "integer"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/Profile"
          }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}